Add tasks to plays to facilitate progress checks
We want to have a way to track the progress learners are making. That's
why a specially-named hidden directory is created at the end of any
playbook, indicating that the playbook has finished successfully.

The only exception is playbook teardown.yml, where at the end of it, all
specially-named hidden directories are deleted.